Jenkins CI Test Report parser
Adds a parser for the Jenkins Test Report json.
Parser gets results for a suite and parses it for the list of TestCaseResultInfo objects.

There are two problems to be solved:
1. configuration
2. repeated names

Ad (1)
TEREC has a concept of test run configuration: same test case being run with different configurations,
for example with different set of libraries or different system setup.

There is some naming convention used in parsing, that case name can include dash (-)
and anything after dash will be considered a configuration.

Configuration must not be None (due to the database model), so if it is not set we use hash (#).

Ad (2)
Jenkins allows to have multiple cases with exactly same class and test case names in the report.
TEREC needs to distinguish those runs. At the moment test run configuration is used.
If subsequent cases with same name are found in the suite results report
then additional # is added to the configuration: so it will be # for 1st occurence, ## for 2nd etc.

TODO: I am not 100% if this is valid approach, maybe instead of using the configuation we should
1. keep the names as they are in jenkins report
2. remove test_configuraton concept completely
3. to solve repeated test names add "-1", "-2" etc?

TODO: what about test case run URL? I think we should add them.