Ignore .idea folder
Some sources recommend to keep part of the folder in VCS,
but some other sources recommend to completely ignore it.

I have decided that in general keeping IDE settings in VCS is a risky way
due to e.g. different (outdated) information that will not be updated
if the maintainers will not use recent version of the IDE.

This change adds .idea to .gitignore.