Add test cases results via api call
A new API call was added to add test cases results to existing suite run.
It makes some validation wrt to existence of the org, project, suite and run.
Then it adds test results to the database.

At this stage we should be able to combine existing API calls to import
some jenkins data from the Cassandra CI.