Read env vars in pipeline commands instead of static
Rationale: to make it testable.

If initialization of the variables such as JENKINS_URL
was done statically then it was pretty inconvenient to mock it
because it was initialized when importing the file with commands.

This change is moving from static initialization to reading
in each pipeline command so that it can be mocked
before calling the command and results are observed.