Use unlogged batches to import test results
Rationale for this change is low performance of importing test results
observed when manually importing data from Cassandra-3.11 runs.

First we are adding pytest-benchmark lib and
a simple benchmark for adding 100 rows via the REST api.

Then we replace stream of unprepared inserts with the UNLOGGED BATCH.

This change improved almost 4x: from 0.9 ops to 3.5 ops.

Note:
this could be potentially improve further
by using prepared statement and execute_concurrent()
but this is not as straightforward as using BatchQuery()

Note:
We may want to split all the data into some fixed-size batches.