Add logic for getting all failed tests for a list of suite runs.
This functionality is required for presenting list of failures
for a suite and a branch. This will be exposed by the CLI command:
failures list --branch main --suite ci

A new module (failures.py) with a new function is added:
get_failed_tests_for_suite_runs()

This function uses execute_concurrent() rather than cqlengine filter
to make it more performant.

A new index on table TEST_CASE_RUN(result) was added.
As it is used in the combination with full partition key
it is always a local (and efficient) call.

To test this new method a new utility - ResultsGenerator - was added.
It allows to generate test suite runs and test case runs data.