Similarity experiments
The goal of those was to find how to automatically detect
same failures based on the stacktrace or error details.

I have asked ChatGPT for the proposal and then tried some of them
in the development/similarity-text.py:
- Levenshtein
- Jaccard
- Cosine on ngrams

Conslusions:
1. it should be probably rejecting: if any ratio is below threshold on any field => not similar
2. we should probably use 2 or 3 different algo
3. possibly ws should lowercase all
4. fancy vector embeddings etc are not required for this problem
5. I have not validated a similar but different stacktrace (sharing part only, but substantial one)

There is also another problem to be solved:
shall we identify test runs or failures in any way?

Apparently we can generate nice hash for the failure using truncated sha-256 on the key fields,
and this hash could be used then for a user e.g. to link issue to it.
Or show detauls of the failure.

There is a design issue as well: should we have a separate "FAILURE" entity?
I am not sure, I need to think about it:
- we would like to detect similar failures
- we would like to create a jira for failure