WIP: Add basic components for finding similar test failures
High Level Design:
==================
Use similarity check on failure fields such as error_stacktrace or stderr
for checking if some failed TestCaseRun is a known failure.

So given some FAIL TestCaseRun
we look for other failures of the same test case (with all configurations)
and then we find for similar failures
building some report including information about:
- totacl checked runs
- passed or skipped test runs
- similar failures
- different failures

This should allow us to return some API Response.

Changes:
========
- adds SimilarityChecker class for two tests check (TODO: performance tuning)
- adds TestCaseRunFailureAnalyser for finding similar test failures in the database (TODO: testing)
- new GET api call /history/...../test-run-check (TODO: test and finish Response building)

Problems:
=========

What should be the response for this call?
Shall we build the history of runs maybe per config in the form of [(build,result)]

Having something like: "config:#" -> PPP_S_FFFxP would be heplful.

There is one catch: we need to do it smart way so that we fill gaps
is configuration or even test case was not run.

This should be possible as we have build numbers.