Initial impl of "tests regression-check" cli.
Current implementation collected check information via REST API
but it is lacking:
1. printing of the history of the new failed tests
2. calculating run_id if not provided (optional, maybe it should be required?)

Side change: TerectCallContext
instead of passing url, org and project this object
can be created and then passed to each api call.

Note:
we will need an api call for this so that it can be used by gh actions for example.