Fix one-suite-at-a-time collection
Solution originally implemented by WindSurf was in many ways incorrect.
It was simply failing as the syntax for "tree" param was incorrect.

I have invested 1h to try and fix it using WindSurf,
but eventually I have fixed it myself in 10min.

Current solution is
- taking number of suites
- taking suites one by one

It has some drawbacks:
- jenkins server connection was closed (16k suites!)
- it is slow as it is not using async while it could
- it opens new requests get each time, it would be better to connect once (?)

So the next step is to send calls in batches of N suites at a moment using async.
Possibly also we should handle connection retry error.