feat: introduce private orgs
This change introduces organisation that are "private"
and access to their projects / test runs should not be public.

This is first step of auth/authz work.

For private organisation tokens are created and stored in db.
Created tokens are returned from PUT /admin/orgs call
together with relevant message.

At this stage tokens are not checked. The work will continue.

Summary of changes:
- new auth component is added with basic token operations
- new entity OrgToken is added to the model
- new field "private" is added to Org model
- when new Org with private=True is created via PUT tokens are generated
- tokens are returned from PUT /admin/orgs endpoint
- tests are adjusted
- tests for auth/tokens.py are added